+++ templates/darray.h
@@ -80,10 +80,8 @@ struct DARRAY_NAME {
 
 #define DARRAY_push DARRAY_IMPL(push)
 #define DARRAY_init DARRAY_IMPL(init)
-#define DARRAY_free DARRAY_IMPL(free)
 #define DARRAY_push_t DARRAY_IMPL(push_t)
 #define DARRAY_init_t DARRAY_IMPL(init_t)
-#define DARRAY_free_t DARRAY_IMPL(free_t)
 
 #ifdef DARRAY_DECLS_ONLY
 
@@ -97,19 +95,11 @@ DARRAY_init(Koliseo* kls);
 
 DARRAY_LINKAGE
 void
-DARRAY_free(DARRAY_NAME* array);
-
-DARRAY_LINKAGE
-void
 DARRAY_push_t(DARRAY_NAME* array, DARRAY_T item);
 
 DARRAY_LINKAGE
 DARRAY_NAME
 DARRAY_init_t(Koliseo_Temp* t_kls);
-
-DARRAY_LINKAGE
-void
-DARRAY_free_t(DARRAY_NAME* array);
 
 #else
 
@@ -156,21 +146,6 @@ DARRAY_init(Koliseo* kls){
 
 DARRAY_LINKAGE
 void
-DARRAY_free(DARRAY_NAME* array){
-    if (array == NULL) {
-        fprintf(stderr,"In %s, at %i: %s(): array was NULL.\n", __FILE__, __LINE__, __func__);
-        exit(EXIT_FAILURE);
-    }
-    if (array->use_temp) {
-        fprintf(stderr, "In %s, at %i: %s(): array uses Koliseo_Temp()\n", __FILE__, __LINE__, __func__);
-        exit(EXIT_FAILURE);
-    }
-    kls_free(array->allocator.kls);
-    return;
-}
-
-DARRAY_LINKAGE
-void
 DARRAY_push_t(DARRAY_NAME* array, DARRAY_T item){
     if (!array->use_temp) {
         fprintf(stderr, "In %s, at %i: %s(): array uses Koliseo\n", __FILE__, __LINE__, __func__);
@@ -211,21 +186,7 @@ DARRAY_init_t(Koliseo_Temp* t_kls){
     return res;
 }
 
-DARRAY_LINKAGE
-void
-DARRAY_free_t(DARRAY_NAME* array){
-    if (array == NULL) {
-        fprintf(stderr,"In %s, at %i: %s(): array was NULL.\n", __FILE__, __LINE__, __func__);
-        exit(EXIT_FAILURE);
-    }
-    if (!array->use_temp) {
-        fprintf(stderr, "In %s, at %i: %s(): array uses Koliseo\n", __FILE__, __LINE__, __func__);
-        exit(EXIT_FAILURE);
-    }
-    kls_temp_end(array->allocator.t_kls);
-    return;
-}
-#endif
+#endif // DARRAY_DECLS_ONLY
 
 // Cleanup
 // These need to be undef'ed so they can be redefined the
@@ -236,11 +197,9 @@ DARRAY_free_t(DARRAY_NAME* array){
 #undef DARRAY_LINKAGE
 #undef DARRAY_push
 #undef DARRAY_init
-#undef DARRAY_free
 #undef DARRAY_push_t
 #undef DARRAY_init_t
-#undef DARRAY_free_t
 #ifdef DARRAY_DECLS_ONLY
 #undef DARRAY_DECLS_ONLY
-#endif // DARRAY_HEADER_H
+#endif // DARRAY_DECLS_ONLY
 #endif // DARRAY_T